IOS-6894 Grooming folders
